+++ installSpigotDependencies.sh
@@ -1,10 +1,23 @@
 #!/usr/bin/env bash
 
+# Print the current Java version for debugging purposes:
+# Java 8: 52 , Java 16: 60
+JAVA_CLASS_VERSION=$(javap -verbose java.lang.Runnable | grep "major version" | cut -d " " -f5)
+echo Current Java class version: $JAVA_CLASS_VERSION
+
 # We only re-build CraftBukkit/Spigot versions that are missing in the Maven cache.
-# Add entries here for every required version of CraftBukkit/Spigot:
+# Add entries here for every required version of CraftBukkit/Spigot.
+
+# The following versions require JDK 8 to build:
+./installJDK.sh 8
+
 if [ ! -f "$HOME/.m2/repository/org/bukkit/craftbukkit/1.14.4-R0.1-SNAPSHOT/craftbukkit-1.14.4-R0.1-SNAPSHOT.jar" ]; then ./installSpigot.sh 1.14.4 ; else echo "Not compiling Spigot 1.14.4 because it is already in our Maven repo" ; fi
 if [ ! -f "$HOME/.m2/repository/org/bukkit/craftbukkit/1.15.2-R0.1-SNAPSHOT/craftbukkit-1.15.2-R0.1-SNAPSHOT.jar" ]; then ./installSpigot.sh 1.15.2 ; else echo "Not compiling Spigot 1.15.2 because it is already in our Maven repo" ; fi
 if [ ! -f "$HOME/.m2/repository/org/bukkit/craftbukkit/1.16.1-R0.1-SNAPSHOT/craftbukkit-1.16.1-R0.1-SNAPSHOT.jar" ]; then ./installSpigot.sh 1.16.1 ; else echo "Not compiling Spigot 1.16.1 because it is already in our Maven repo" ; fi
 if [ ! -f "$HOME/.m2/repository/org/bukkit/craftbukkit/1.16.2-R0.1-SNAPSHOT/craftbukkit-1.16.2-R0.1-SNAPSHOT.jar" ]; then ./installSpigot.sh 1.16.2 ; else echo "Not compiling Spigot 1.16.2 because it is already in our Maven repo" ; fi
 if [ ! -f "$HOME/.m2/repository/org/bukkit/craftbukkit/1.16.4-R0.1-SNAPSHOT/craftbukkit-1.16.4-R0.1-SNAPSHOT.jar" ]; then ./installSpigot.sh 1.16.4 ; else echo "Not compiling Spigot 1.16.4 because it is already in our Maven repo" ; fi
+
+# The following versions require JDK 16 to build:
+./installJDK.sh 16
+
 if [ ! -f "$HOME/.m2/repository/org/bukkit/craftbukkit/1.17-R0.1-SNAPSHOT/craftbukkit-1.17-R0.1-SNAPSHOT-remapped-mojang.jar" ]; then ./installSpigot.sh 1.17 ; else echo "Not compiling Spigot 1.17 because it is already in our Maven repo" ; fi